refactor: move MarkdownProcessor into support package
